pyproject.toml
[tool.poetry]
name = "nonebot_plugin_roll"
version = "0.2.2a1"
description = "Roll a dice!"
authors = ["KafCoppelia <k740677208@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_roll"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b3"
nonebot-adapter-onebot = "^2.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tencentyun"
url = "https://mirrors.tencent.com/pypi/simple"
default = true
